% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsset.R
\name{tsset}
\alias{tsset}
\title{Title Time Series Set}
\usage{
tsset(df, start = c(1, 1), frequency = "daily")
}
\arguments{
\item{df}{A data frame. The data set that the function will operate on.}

\item{start}{A numeric value. The starting point or index}

\item{frequency}{A numeric value. The frequency of observations (e.g., for time series).}
}
\value{
The output of the function, such as a data frame, vector, or plot.
}
\description{
Title Time Series Set
}
